refactor(site): decompose AgentsEmptyState + fix state-sync effects
Extract ConfigureAgentsDialog to its own file, removing ~210 lines from
AgentsEmptyState (429 → 224 lines). The dialog now manages its own tab
navigation state (configureSectionOptions, activeConfigureSection) and
icon imports internally.

Key changes:
- Remove dead canUseLocalWorkspaceMode code paths (hardcoded false):
  eliminated selectedWorkspaceMode state, selectedWorkspaceModeRef,
  localWorkspaceValue, the local-workspace Select option, and
  workspace_mode from CreateChatOptions.
- Replace 3 state-sync useEffects with derived state:
  1. selectedModel validated against modelOptions every render
  2. activeConfigureSection validated inside ConfigureAgentsDialog
  3. canUseLocalWorkspaceMode guard removed (dead code)
- Extract ConfigureAgentsDialog to ConfigureAgentsDialog.tsx (213 lines)
  with a clean props interface for system prompt management.
- Clean up 10+ imports from AgentsPage.tsx (Dialog components,
  ScrollArea, TextareaAutosize, icon components, ChatModelAdminPanel).

Net result: 3 useEffects eliminated, 2 useStates removed, 2 useRefs
removed, 1 prop removed, AgentsEmptyState well under 250 lines.
